Added unit test video games controller

+++ src/main/java/com/rogelio/basecamp/TrackMAPI/videogame/VideoGamesServiceImplementation.java
@@ -1,5 +1,7 @@
 package com.rogelio.basecamp.TrackMAPI.videogame;
 
+import com.rogelio.basecamp.TrackMAPI.errorhandling.BadRequestException;
+import com.rogelio.basecamp.TrackMAPI.errorhandling.NoContentException;
 import com.rogelio.basecamp.TrackMAPI.errorhandling.RecordNotFoundException;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,37 +18,64 @@ public class VideoGamesServiceImplementation implements VideoGamesService{
     public VideoGamesServiceImplementation(){ }
 
     @Override
-    public void createGame(VideoGame game) {
-        gamesRepository.save(game);
+    public VideoGame createVideoGame(VideoGame videoGame) {
+        return gamesRepository.save(videoGame);
     }
 
     @Override
-    public List<VideoGame> getAllGames() {
+    public List<VideoGame> getAllVideoGames() {
         return gamesRepository.findAll();
     }
 
     @Override
-    public VideoGame getGame(ObjectId gameId) {
-        return gamesRepository.findById(gameId).orElseThrow(() -> new RecordNotFoundException("Can't find " + gameId.toHexString() + ". It does not exist"));
-    }
-
-    @Override
-    public VideoGame updateUser(ObjectId gameId, VideoGame game) {
-        if(!gamesRepository.existsById(gameId)){
-            throw new RecordNotFoundException("Can't find " + gameId.toHexString() + ". It does not exist");
+    public VideoGame getVideoGame(String gameId) {
+        //if hex string is invalid, throws bad syntax exception
+        if(!ObjectId.isValid(gameId)){
+            throw new BadRequestException("Invalid Id: " + gameId);
         }
 
-        game.setGameId(gameId);
-        return gamesRepository.save(game);
+        ObjectId objectId = new ObjectId(gameId);
+
+        return gamesRepository.findById(objectId).orElseThrow(() -> new RecordNotFoundException("Can't find " + objectId.toHexString() + ". It does not exist"));
     }
 
     @Override
-    public void deleteGame(ObjectId gameId) {
-        gamesRepository.deleteById(gameId);
+    public VideoGame updateVideoGame(String gameId, VideoGame videoGame) {
+        if(!ObjectId.isValid(gameId)){
+            throw new BadRequestException("Invalid id: " + gameId);
+        }
+
+        ObjectId objectId = new ObjectId(gameId);
+
+        if(!gamesRepository.existsById(objectId)){
+            throw new RecordNotFoundException("Can't find " + objectId.toHexString() + ". It does not exist");
+        }
+
+        videoGame.setGameId(objectId);
+        return gamesRepository.save(videoGame);
     }
 
     @Override
-    public void deleteAllGame() {
+    public String deleteVideoGame(String gameId) {
+        if(!ObjectId.isValid(gameId)){
+            throw new BadRequestException("Invalid Id: " + gameId);
+        }
+
+        ObjectId objectId = new ObjectId(gameId);
+
+        if(!gamesRepository.existsById(objectId)){
+            throw new RecordNotFoundException("Can't find " + objectId.toHexString() + ". It does not exist");
+        }
+        gamesRepository.deleteById(objectId);
+        return "Succesfully deleted video game";
+    }
+
+    @Override
+    public String deleteAllVideoGames() {
+        if(gamesRepository.findAll().isEmpty()){
+            throw new NoContentException("The server successfully processed the request, but is not returning any content");
+        }
         gamesRepository.deleteAll();
+        return "Succesfully deleted all video games";
     }
 }
